Moved dialog windows to nldialogs
