+++ app/nldialogs/patchsummary.go
@@ -1,4 +1,4 @@
-package nlwidgets
+package nldialogs
 
 import (
 	"sync"
